requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.10.0
google-generativeai==0.3.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# Vector Database
pinecone-client==3.0.2
chromadb==0.4.22

# Data Processing
pandas==2.2.0
numpy==1.26.3
lxml==5.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
